Consolidate ROM integration test binaries
This puts all the ROM integration tests into a single binary named
rom_integration_tests-xxxx, which makes it much easier to run all the
tests on a remote system like a FPGA dev board.

This also has the side benefit of reducing compilation time and reducing
the total size of all test binaries, leading to faster transfers over
the network.