FPGA CI: Fix NTP clock sync, panic-after-softlock, and systemd init order (#1277)
* fpga-image: Use softlock_panic sysctl correctly.

Previously, we were setting this as if it was a timeout value, but it's
actually just a boolean.

See
https://www.infradead.org/~mchehab/kernel_docs/admin-guide/sysctl/kernel.html#softlockup-panic

* fpga-image: Run startup script after everything else.

Before this change, systemd hadn't fully initialized when the startup
script was run (as it was blocking continued startup).

* fpga-image: Synchronize clock via NTP.

GitHub rejects runners with a local clock that varies by more than 15
minutes. Unfortunately, the RTC on the zcu104 boards is terrible,
drifting more than 30 seconds a day.