Reduce core count for FPGA firmware-build job. (#1413)
We'll soon be launching 8 instances of this job simultaneously as part
of the nightly release. Unfortunately, at 32-cores each this will blow
our GCE quota in us-central, and GCE is rejecting my requests for more
quota. So for now, we'll decrease to 16 cores, at least until we can
improve the VM creation logic to retry after quota failures.