Ci cherry pick 12 04 (#2891)
* Update GitHub runner function infra

* Update dependencies
* Bump Go Runtime
* Bump Go toolchain

* ci-tools: Add rtool subcommand to download artifact (#1643)

This is useful to keep the runner images up-to-date.

* Add go.sum (#2220)

This makes `rtool` easier to build with nix.

* Add VCK-190 tools disk to github-runner

* Add FPGA Boss error condition check

If the github runner has an uncaught exception the FPGA needs a reset to
continue.

* ci/fpga-boss: add SDMux implementation for usbsdmux (#2143)

* ci/fpga-boss: add SDMux implementation for usbsdmux

Signed-off-by: leongross <leon.gross@9elements.com>

* Add flake.nix to FPGA Boss and rtool (#2283)

This makes it easier to use these CI tools in Nix.

* Add instructions for FPGA Host Runner (#2236)

* Add instructions for FPGA Host Runner

* Apply suggestions from code review

Co-authored-by: Zach Halvorsen <zhalvorsen@google.com>

---------

Co-authored-by: Zach Halvorsen <zhalvorsen@google.com>

* Increase disk size when using VCK190 tools (#2300)

* Check for Bus Error in FPGA Boss (#2307)

* rtool: Add branch to artifact download (#2305)

It's useful to be able to download the artifact from non-main branches
for testing. Additionally, the VCK-190 image build will (most-likely) be
published to a difference branch.

* Update NixOS config for hostrunners (#2324)

* Add zhalvorsen ssh key
* Add systemd timer to fetch latest fpga image
* Get rtool from flake

* Increase runner disk size to 32 GB (#2335)

Need more space to cross-compile caliptra-mcu-sw

* Hostrunner update (#2339)

* Rotate clundin key
* Disable VCK-2,3,5 jobs
* Make FPGA jobs user jobs to avoid using sudo.
* Add some helper scripts

* Add Docker container to host caliptra build tools (#2372)

This will be used particularly in caliptra-mcu-sw as well as local
cross-compilation. Eventually we will
want to separate this to an infra specific repo.

* Sync host runner changes (#2406)

* [fpga-boss] add command to reset FPGA from FTDI chip (#2426)

* [fpga-boss] fix comment

Signed-off-by: Tim Trippel <ttrippel@google.com>

* [fpga-boss] make `sdwire` and `usbsdmux` nonmandatory args

Not all subcommands require these args to be set. Refactor the code such
that only subcommands that use these args require them.

Signed-off-by: Tim Trippel <ttrippel@google.com>

* [fpga-boss] add command to reset FPGA from FTDI chip

Currently the only way to reset the FPGA board was to also toggle the SD
card mux which, as a byproduct, also toggled the FPGA reset. This adds a
dedicated FPGA reset command to fpga-boss.

Signed-off-by: Tim Trippel <ttrippel@google.com>

---------

Signed-off-by: Tim Trippel <ttrippel@google.com>

* nit: fix ci-tools README link (#2436)

Signed-off-by: leongross <leon.gross@9elements.com>

* Fix weekly image download cron job for 2.x images (#2438)

Need to use the full path for unzip when running a systemd job.

* Add bitstream bucket to GitHub CI (#2439)

* Add Caliptra Core 2.x compiler tools (#2387)

* [fpga_boss] Check the file was flashed to the SD card correctly. (#2451)

This adds a check after copying the image to the SD card to check it was written correctly.

* ci-tools: add required (stateVersion)["https://nixos.org/manual/nixos/stable/options.html#opt-system.stateVersion"] (#2469)

Signed-off-by: leongross <leon.gross@9elements.com>

* add hostrunner bo-0 (#2437)

* ci-tools/hostrunner: add usbsdmux

Signed-off-by: leongross <leon.gross@9elements.com>

* ci-tools/host-runner: add Bochum-0 runner

Signed-off-by: leongross <leon.gross@9elements.com>

* fix: adjust secrets path to current default nixosSystem.specialArgs.user

Signed-off-by: leongross <leon.gross@9elements.com>

* ci: update fpga-boss.nix to support usbsdmux

Signed-off-by: leongross <leon.gross@9elements.com>

---------

Signed-off-by: leongross <leon.gross@9elements.com>

* ci-tools/host-runner: add user lingering (#2495)

Signed-off-by: leongross <leon.gross@9elements.com>

* [build-image] Use default rust profile instead of minimal

This makes the image larger but faster to use. With the minimal profile
all the compiler tools need to be downloaded at each invocation.

* [fpga-boss] Fix SDMUX usb path

* [hostrunner] - Pull FPGA boss in with a flake (#2537)

This makes updating FPGA boss as easy as `nix flake update`.

* [ci-tools]: Remove Bitstream bucket delete lifecycle

Let's keep object indefinitely until we have a plan for generating new
bitstreams.

* [hostrunner]: Run segmented FPGA on kir-0

* Onboard VCK-6

* [hostrunner]: Save old fpga images and add a clean up job

This makes sure we don't overwrite a rollback image.
Also clean the images up after 4 weeks so we don't fill our disk with
old images.

* Add support for a bigger GCE instance disk (#2599)

* [Hostrunner]: FPGA image changes

* Run segmented image on VCK-2 and VCK-6
* Download 2.x images from weekly cron stored on main

* [hostrunner]: Few small updates (#2629)

* Add kir-7 FPGA
* Download segmented FPGA image
* Use image from cron job in core runners

* Run subsystem-2.0 on VCK-7

* Hostrunner tweaks (#2691)

* Run autoformatter

* Split SSH keys into a separate config file

* Add instructions for code formatting

* [FPGA-BOSS]: Reduce job timeout

Reduce job timeout from 2 hours to 1 hour. A happy test run should take
around 45 minutes. An FPGA that crashes or has failures will be stuck
for 2 hours until this reset is triggered. 

Reducing the timeout should improve FPGA availability, since they will
have more uptime.

* hostrunner: Update nix flakes (#2732)

* Integrate's Caliptra FPGA boss timeout decrease

* FPGA Boss: Bump timeout

Bump FPGA job timeout back to 3 hours because the subsystem job takes a
long time.

* Increase boot timeout (#2795)

#2794 add's a sleep to the boot configuration, to allow the NTP service to update.

Increase boot timeout for now, until a more permanent fix is enabled.

* CI FPGAs: Re-allocate FPGAs (#2815)

Also, update nix flake to include timeouts

* hostrunner: Segmented subsystem (#2873)

Migrate FPGAs to segmented subsystem.

* Fix FPGA Boss clippy from newer compiler

---------

Signed-off-by: leongross <leon.gross@9elements.com>
Signed-off-by: Tim Trippel <ttrippel@google.com>
Co-authored-by: Kor Nielsen <kor@google.com>
Co-authored-by: leongross <leon.gross@9elements.com>
Co-authored-by: Zach Halvorsen <zhalvorsen@google.com>
Co-authored-by: Timothy Trippel <5633066+timothytrippel@users.noreply.github.com>