Cherry pick 12 08 (#2932)
* Fix `test_cryptographic_mailbox::test_derive_stable_key_from_rom` (#2836)

The tests that stop at ROM don't include a Core FW image, but do include manifest and MCU FW so it was failing while trying to unwrap the Core FW image. This will now return from the subsystem `boot` function if just the Core FW image is excluded.

* Always scramble FPGA MAC address (#2874)

Since the FPGAs share a MAC address, they will step over each other if
on the same network.

* hw-model: Split OTP in two we can use half for DOT flash (#2872)

The OTP memory is 16384 bytes, but only about ~3300 bytes are used at
the moment for the default fuse map.

We need a simulated flash or NV memory for Device Ownership Transfer in
MCU. It is much easier to use some of the unused space in the OTP memory
than it would be to add another BRAM in the FPGA, wire it up, etc.

So, we limit the OTP space to be only 8192 bytes and reserve the other
8192 bytes for a simulated "flash" storage. (Mostly we need to prevent
the OTP provisioning from clearing this storage, so we limit the
otp_slice to 8192 bytes.)

* [fpga-image] add developer ssh keys to dev image build (#2861)

This adds developer ssh keys (for cfrantz and moidx) to the FPGA dev
image build. Additionally, this alphabetizes the ordering of all
existing SSH keys.

Signed-off-by: Tim Trippel <ttrippel@google.com>

* fix `test_invoke_dpe::test_export_cdi_destroyed_root_context` (#2838)

This test gets the fatal error while still coming up from warm reset. This needed warm_reset in subsystem to return before the full flow is complete so we can catch the error. This is changed to wait until MCU ROM sets the fuses in warm reset. The `fw_flow_status` error is cleared on warm reset so it shouldn't immediately return from the milestone set during cold reset.

This also overrides the `warm_reset_flow` HwModel function in subsystem because those registers are supposed to be set by MCU ROM instead of the HwModel.

* modify preamble and description to update LMS and sig with MLDSA87 and sig (#2839)

* modify preamble and description to update LMS and sig with MLDSA87 keys and signatures,  the sizes is 2592 bytes for the public key and 4628 bytes for the signature.

* we should keep both LMS and MLDSA,  not replace LSM with MLDSA.  udpate to keep both

* code review feedback to take out field not in AuthManifestImageMetadata structure.

* missed one

* [doc] Updating ROM README (#2914)

* Update MCU Commit

* Test fixes

* Filter out FE UDS tests for now. Should be solved by https://github.com/chipsalliance/caliptra-sw/pull/2936
* Increase invoke_dpe step count to 1000.
* Filter out OCP LOCK tests that need to write fuses. Will be resolved
  by https://github.com/chipsalliance/caliptra-sw/pull/2915.

* Disable OCP LOCK Runtime tests

Will be enabled in https://github.com/chipsalliance/caliptra-sw/pull/2915.

---------

Signed-off-by: Tim Trippel <ttrippel@google.com>
Co-authored-by: Zach Halvorsen <zhalvorsen@google.com>
Co-authored-by: Christopher Swenson <chswenson@microsoft.com>
Co-authored-by: Timothy Trippel <5633066+timothytrippel@users.noreply.github.com>
Co-authored-by: meilu-git <mei.lu@businessintegra.com>
Co-authored-by: Vishal Mhatre <38512878+mhatrevi@users.noreply.github.com>