Add more soc_ifc boot settings from MCU ROM (#2936)
* fpga_subsystem: Set initial UDS and FE fuses via OTP fuses

This will make tests succeed that need UDS and FE to be programmed to a
certain value to have a predictable IDEVID key. This are for instance
the smoke_tests.

Signed-off-by: Arthur Heymans <arthur.heymans@9elements.com>

* fpga-realtime: Use fuses parameters for UDS seed and FE

Replace hardcoded DEFAULT_UDS_SEED and DEFAULT_FIELD_ENTROPY values with
values from the fuses parameters when initializing the FPGA wrapper.
This allows proper configuration of security-critical values based on
the provided parameters rather than relying on defaults.

It does not look we need other values than default in there but it's
still a good idea.

Signed-off-by: Arthur Heymans <arthur.heymans@9elements.com>

* Fix test_image_validation::cert_test_with_ueid

Set the initial cptra_dbg_manuf_service_reg via InitParams and properly
set idevid_cert_attr.

Signed-off-by: Arthur Heymans <arthur.heymans@9elements.com>

---------

Signed-off-by: Arthur Heymans <arthur.heymans@9elements.com>